Add arbitrary hooks to the renderer.
Arbitrary hooks in the renderer allow any function with the @draw_hooks
decorator to have a list of pre and post functions to be called. This is
intended to make stuff really easy to override, rather than obeying Python's
class override rules (this way is more flexible).

As a proof of concept, the reader's various abilities have been moved to this
new system of hooks by default.

Signed-off-by: Jack Miller <jack@codezen.org>