First Draft: Threading / Partial Update
This should increase Canto performance a lot. First, the interface updates
incrementally, both in terms of feeds and in terms of the amount of information
given from each feed. In addition, the entire update process has been moved into
threads and out of the main input / graphical update path so the interface
doesn't go non-responsive when Canto gets choked by large (i.e. initial)
updates, despite the new partial update strategy.

There are, however, still some outstanding issues with this code, like what to
do with the current selection after a filter change and when to clear the queue
of updating threads.