Redone threading, draft 2
This commit moves more of the tag work into the background thread, which has
been consolidated and revamped to communicate with the main thread with two
Queues, update, and a updated. The main thread puts a tuple on the update queue,
the thread reads it, processes it and puts the result on the updated queue where
the main thread then performs the update when there's no input.