Set proper default triggers, hook up config
To use triggers, just add strings to the triggers array. For example, remove
"interval" trigger (i.e. the default timed one) and add "change_tag" (updated
when you move to another tag).

    triggers.remove("interval")
    triggers.append("change_tag")

The only other trigger is "signal", which allows you to issue a SIGUSR1 to the
canto instance to start an update. And of course you can do the usual C^r to
manually force an update.