Change semantics for c-f keeping items
From now on setting `keep` for a feed below the number of items in that feed's
XML will have no effect. So, if keep = 40 (the default), but your feed includes
85 items (i.e. Washington Post), you're still going to get all 85 items
recorded.

Functionality lost by this can be (better) implemented with a `show` variable
that limits the number of items shown in the interface, but the state for all
items is still kept on disk.