Try to avoid too long update queues
Particularly with the "change_tag" trigger, the update queue can explode to
large sizes which causes put() to block and CPU to shoot up for no reason
(usually), so if the update queue has items on it, update will do nothing.

More intensive operations (like filter changes) will flush the queues before
proceeding so those are untouched and the regular updates are processed fast
enough that if there are still updates on the queue, a previous update was
invoked recently enough that it's pointless.

This might need to be supplemented with some heuristic like if the size of the
queue is < half the number of tags to be updated, then go ahead and continue but
I think this isn't necessary.