First draft at modular config.
There are some problems with the code as the client is fucked up compared to the
vanilla experimental branch. However, the motivation behind the modularization
is to validate and test small portions of the config more easily so in the
process of writing the validations and tests, I think these will be worked out.
Until then, this is unsuitable to be merged into experimental.