Make the tags variable set to a more sane default
Now tags sets itself to have more than the default tags in it by default. So if
you define three feeds, add the "news" tag to the first two, and the "art" tag
to the third, tags is automatically set to

tags = [ ["Feed 1", "Feed 2", "Feed 3"], ["news"], ["art"] ]

instead of

tags = [ ["Feed 1", "Feed 2", "Feed 3"] ]

The tags variable is still manipulatable if you want different tag behavior, but
it should largely be unnecessary after this point.