Switch to custom Queue
I hate to do this, but the builtin Python Queue class fails for one important
thing: without maintaining a synchronized list of objects in the Queue, it
doesn't appear to have a method to return whether an object is in the queue or
not. As such, I've implemented QueueList which is similar to Queue in that it's
functions are named the same and behave similarly, but there are no timeouts and
duplicate objects are not allowed.

The basic advantage of this is that if I do something like switch a filter and
queue up the resulting work for the thread, and then later queue up an update or
another filter change if the original work item hasn't been processed, there
won't be two identical work items when one will do.