Initial import of c-i
Canto-inspect (c-i) is pretty basic at the moment, just taking

    canto-inspect URL

and nothing else. It is, however, entirely functional for what I want it to do,
so it's possible that this will be the released version. Later, I'd like to
integrate it with canto's config so that `canto-inspect MAINTAG` will work from
disk, or so that scripts can work (although, you could pipe the script into a
file and use a file:// URL to get around that one).