Initial process based draft.
- Add python-processing dependency
- Moves away from horrendously broken (http://www.dabeaz.com/python/GIL.pdf)
threads and to processes.

I'm not happy with this yet, but the performance and responsiveness seem to have
enjoyed quite a nice boost now that we have escaped the GIL.

Basic overview: instead of a thread, a new process is forked that grabs all of
the filter and sort lambdas out of the previous context as well as the list of
feeds. Two pipes are open to it, one to the process and one from the process.
Then a protocol of pickle-able objects is passed back and forth through the
pipes.

I attempted to use the python-processing Queue and Process classes, but I
couldn't get them to cleanup properly because the Queues seem to get out of sync
with themselves and throw exceptions trying to wait on processes that no longer
exist. Fortunately, the system only has two nodes so Pipes function well enough
in place of the Queues since there's only one worker process.

There are some issues remaining:
   - Filters and sorts are collected at configuration time and can only be
   passed into the worker thread at its initial calling. This works
   transparently for filters and sorts defined in the usual places (i.e. feeds
   tags, defaults), but set_* functions that take an argument of a sort and
   return a keybind need to register the filters/sorts so they exist in the
   worker thread. This is going to be tricky.

      - Either, a global register_filter/register_sort function needs to exist
      - or the gui needs to be able to amend the list and restart the process.

   - The process code is untested, a problem I hope to solve by pushing this
   into the repo.