Fix setting sorts / filters by keybind
The solution is transparent to the end user. However, if you're going to write
any code that's somehow going to introduce a new filter / sort without it being
in a usual place (i.e. in a global or tag definition) then you need to:

   from cfg.filters import all_filters
   from.cfg.sorts import all_sorts

And make sure that it's included in there or you're going to cause trouble when
you attempt to use your code. Of course 99% of people won't care about this.