Fix uniprocessors
If you want to hear a bug war-story, feel free to ask about this one in #canto.
The short version is that for some, completely stupid and unknown reason,
socket.setdefaulttimeout() affects pipes. Yep.

Apparently the core/core2s I test on regularly are fast enough to avoid this all
together? I don't know. All I know is that turning off the timeout fixes the
issue.