Use indices in tag diffs.
This fixes (I think) the dichotomy between items in the gui and items in the
feeds. In addition to merging the items in the feeds, the story objects are
translated from the merged feeds rather than freshly created in the worker
process.

This is relatively untested (but practically all of the process based stuff is)
but it fixes core behavior.