Make a default no-content stub.
This is written to the feed file if the file doesn't exist. It's a valid feed
dict with one story saying to read the fetchlog for errors. Most likely, you'll
only ever see this if the feed's glitched, but it'll keep canto from attempting
to refetch glitched feeds over and over since it doesn't put any disk presence
down.