Intelligent precaching
This patch makes it so that filters and sorts can request to precache the
content they use. For example, the by_date sort requests the "updated_parsed"
key to be cached so that for every comparison done, the updated_parsed key
doesn't have to be fetched from disk (which is extremely expensive, especially
when you're sorting a couple of hundred of stories).