Restart worker on unknown filter
Generally, we like to know about all the filters after config validation time.
Unfortunately, to use really dynamic filters (like with search_filter), we have
to be flexible. This makes it so that filters that aren't resident in the worker
process (i.e. not in cfg.all_{sorts,filters}) cause a restart.