Stop enforcing rate in canto
This is a stupid bug but for most people it has little appreciable effect
because the default rates are so low. If c-f fetches the feed every n minutes
and canto only reads it every n minutes, the worst case scenario is that you're
lagging a full n minutes behind the real feed (roughly, and only if you're
perfectly out of sync).

As I said, with the default rate of 5, this is usually no big deal. However, now
that it's apparent that people in the wild are using 60 minute keeps and higher
(yikes!), that's extremely bad behavior.