Add never_discard
Add never_discard, which allows the user to keep c-f from eliminating items with
any tag. It could be used for a number of applications, but
never_discard("unread") is likely to be the widest use. Even though "unread"
isn't a tag (just "read"), it's a special case that's negated. If someone
complains, the syntax can also be expanded to allow never_discard(!tag) with
some sugar for unread == !read. I don't think this is necessary.

You can use this like so:

never_discard("unread")