Add wget_link reader_key
A nice little helper to download content into a directory. You use it like so:

reader_keys['w'] = wget_link('/path/to/directory')

It runs in the background, of course, but as of yet there's no way to tell when
it's actually done downloading. On the plus side though, it doesn't steal the
terminal or stop canto.