Make item in reader immune from retraction
This fixes the problem of items disappearing and causing issues with using the
'n'/'p' binds in the reader. Before, the item would disappear, selecting a new
item behind the scenes (which works, if you exit the reader and see that it's
selected), then hit 'n' which would take you to the next item, rather than the
correct item.

I thought that I had fixed this by making the item set unread *after* the reader
was closed, but that only fixes show_unread (even though that's the most popular
filter) and doesn't take into account the idea that the item might be read
*before* you start the reader anyway (i.e. you decided to revisit a read
article, or you accidentally closed the reader, etc.)