Potential fix for issue 3 (long running crash)
I'm not 100% sure because it was speculative in the first place, but my two
faithful testers jaacoppi and evg_krsk both seemed to have their problems
alleviated today.

In this case, it could be my fault for SIG_IGN signals that really just needed
to be SIG_DFL'd (i.e. unlike the other SIGCHLD problem in which I had a custom
handler to use instead).

Signed-off-by: Jack Miller <jack@codezen.org>