Work harder to maintain cursor through updates
So now, in addition to the cursor improvements, whenever the gui is updated, try
harder to stay close to where you were after a refilter / update.

Sadly, the nature of Canto's interface (which I'm not going to change) doesn't
allow for 100% certainty that the cursor will stay nailed to the middle, but we
try to behave.

Signed-off-by: Jack Miller <jack@codezen.org>